Sandbox: ICP: Fix distance map version: 0.5s for 01.pbm over 02.pbm.
	* mln/algebra/vec.hh: Add check for dimension in to_point().
	* sandbox/jardonnet/test/plotscript: Add gnuplot script.
	* sandbox/jardonnet/test/Makefile: Add g++ flags: Unexpected behavior disappear with -ffloat-store (float are not stored in register).
	* sandbox/jardonnet/test/check: Update test script.
	* sandbox/jardonnet/registration/quat7.hh: .
	* sandbox/jardonnet/registration/icp.hh: .
	* sandbox/jardonnet/registration/projection.hh: .
	* sandbox/jardonnet/registration/tools.hh: .


git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@1819 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/sandbox/jardonnet/registration/quat7.hh
@@ -99,7 +99,7 @@ namespace mln
     // qR
 
     algebra::mat<P::dim,P::dim,float> Mk;
-    for (unsigned i = 0; i < C.npoints(); ++i)
+    for (size_t i = 0; i < C.npoints(); ++i)
       {
         algebra::vec<P::dim,float> Ci = C[i];
         algebra::vec<P::dim,float> Xki = Xk[i];
